--- cw/cassandra.py
@@ -1,2 +1,0 @@
-#from cassandra.cluster import Cluster
-#cluster = Cluster()
